server/requirements.txt
Flask
RPi.GPIO
gTTS
mpyg321
wikipedia
PyYAML